backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
aiosqlite==0.20.0
pydantic>=2.11
python-dotenv==1.0.1
PyNaCl==1.5.0
base58==2.1.1
websockets==12.0
python-multipart==0.0.9
groq>=0.9.0
solana>=0.32.0
solders>=0.20.0
